Fix indentation of pipeline.close() and main call
Earlier versions do not run unless called with the -v options since the main
call was indented under `if verbose:`.
This also solves issue https://bitbucket.org/psitomcat/dpc_reconstruction/issue/2/logerror-not-shown
the logs were not being shown because the program would not actually run in
the first place.

The pipeline.close() call is now in the outer indentation level so that the
app can go through all the intended loops.